REST API helpers file moved to directory /server
